pyproject.toml
[project]
name = "refine-lab"
version = "0.1.0"
description = "RV-GRPO: Rubric-Verifiable GRPO for Behavioral Alignment in Small Language Models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.12"
authors = [
    {name = "Jkanishkha0305"},
]

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "transformers>=4.38.0",
    "datasets>=2.16.0",
    "tokenizers>=0.15.0",
    "accelerate>=0.27.0",

    # Fine-tuning
    "peft>=0.8.0",
    "bitsandbytes>=0.42.0",
    "trl>=0.14.0",

    # Experiment tracking
    "wandb>=0.16.0",

    # Evaluation
    "evaluate>=0.4.0",
    "scikit-learn>=1.3.0",

    # Data
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
# Flash attention (needs separate install on GPU machines)
flash = ["flash-attn>=2.5.0", "packaging", "ninja"]

# Quantization & deployment
deploy = ["auto-gptq>=0.7.0", "llama-cpp-python>=0.2.0", "optimum>=1.16.0"]

# External benchmark evaluation
eval = ["openai>=1.0.0", "anthropic>=0.18.0"]

# Development
dev = ["pytest>=7.4.0", "ruff>=0.1.9", "black>=23.12.0"]

# Full install (everything)
all = ["refine-lab[flash,deploy,eval,dev]"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["data*", "experiments*", "evaluation*", "deployment*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
